[Refactor] Member serviceImpl로 리팩토링
